+++ CHANGELOG.md
@@ -1,3 +1,25 @@
 # ChangeLog
 
-# 0.1.0
+## 0.1.2
+
+2024.01.07
+
+- added vectorization from dict of scores
+- added priority fusion
+
+## 0.1.1
+
+2024.01.06
+
+- added vector based selection and re-ranking
+
+
+## 0.1.0
+
+2024.01.04
+
+Initial release.
+
+- `basic`, `voter` and `borda` (simple and tournament-style) decision fusion methods.
+- support function for vector scaling (`scale`).
+- ranking functions (`rank`)